backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
beautifulsoup4
sentence-transformers
faiss-cpu
lxml
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface